==> notebooks/01_api_exploration.ipynb <==
import requests
import pandas as pd
import json

url = "https://api.worldbank.org/v2/country/USA;CHN;DEU;MEX;IND;VNM/indicator/NE.EXP.GNFS.CD"

params = {'date': '2020:2024',
    'format': 'json',
    'per_page': 1000}

response = requests.get(url, params=params)
data = response.json()

print(json.dumps(data[1][0], indent = 2))

records = []

for item in data[1]:

    records.append({'country': item['country']['value'],
        'country_code': item['countryiso3code'],
        'year': item['date'],
        'value': item['value'],
        'indicator': item['indicator']['value']
    })

df = pd.DataFrame(records)
print(df.head())
